refactor(github-action): refactored github-action folder
