+++ src/github-action/index.ts
@@ -1,5 +1,5 @@
-export * from './commitlint';
-export * from './npm';
+export * from './config/commitlint';
+export * from './config/npm';
 export * from './project';
 export * from './typescript';
-export * from './samplecode';
+export * from './config/samplecode';
